# Use io.StringWriter