Benchmarks against the previous implementation
[new-iowriter.txt](https://github.com/choonkeat/dom-go/files/13308575/new-iowriter.txt)
[new-stringwriter.txt](https://github.com/choonkeat/dom-go/files/13308577/new-stringwriter.txt)
[old.txt](https://github.com/choonkeat/dom-go/files/13308579/old.txt)


```
benchstat old=old.txt new=new-stringwriter.txt  
goos: darwin
goarch: arm64
pkg: github.com/choonkeat/dom-go
               │     old     │                 new                 │
               │   sec/op    │   sec/op     vs base                │
AttrHTML-8       123.9n ± 2%   157.5n ± 1%  +27.12% (p=0.000 n=10)
NodeHTML-8       520.4n ± 0%   670.9n ± 2%  +28.91% (p=0.000 n=10)
HtmlTemplate-8   3.312µ ± 1%   3.378µ ± 1%   +1.98% (p=0.001 n=10)
geomean          597.8n        709.3n       +18.67%

           │     old     │                 new                 │
           │  AttrHTML   │  AttrHTML    vs base                │
AttrHTML-8   9.479M ± 4%   7.499M ± 7%  -20.89% (p=0.000 n=10)

               │     old      │                  new                   │
               │     B/op     │     B/op      vs base                  │
AttrHTML-8         128.0 ± 0%     160.0 ± 0%  +25.00% (p=0.000 n=10)
NodeHTML-8         456.0 ± 0%     800.0 ± 0%  +75.44% (p=0.000 n=10)
HtmlTemplate-8   1.352Ki ± 0%   1.352Ki ± 0%        ~ (p=1.000 n=10) ¹
geomean            432.3          561.6       +29.92%
¹ all samples are equal

               │    old     │                 new                  │
               │ allocs/op  │ allocs/op   vs base                  │
AttrHTML-8       4.000 ± 0%   5.000 ± 0%  +25.00% (p=0.000 n=10)
NodeHTML-8       11.00 ± 0%   16.00 ± 0%  +45.45% (p=0.000 n=10)
HtmlTemplate-8   53.00 ± 0%   53.00 ± 0%        ~ (p=1.000 n=10) ¹
geomean          13.26        16.19       +22.05%
¹ all samples are equal

           │     old     │                 new                 │
           │  NodeHTML   │  NodeHTML    vs base                │
NodeHTML-8   2.297M ± 1%   1.797M ± 3%  -21.77% (p=0.000 n=10)

               │     old      │              new               │
               │ HTMLTemplate │ HTMLTemplate  vs base          │
HtmlTemplate-8    354.2k ± 0%    351.9k ± 2%  ~ (p=0.631 n=10)
```