@choonkeat 2023-11-09
the numbers don't look good, huh :-(

my understanding is that the goal is to get `func (e Node) Write(w io.StringWriter)` where we can supply our own `io.StringWriter`

how about we instead expose our internal helper

```diff
-func (a Attribute) buildHTML(sb *strings.Builder) *strings.Builder {
+func (a Attribute) HTMLBuilder(sb *strings.Builder) *strings.Builder {
```

so instead of _only_ being able to call `HTML()`, now we have the choice to provide our `strings.Builder`