@choonkeat 2023-11-09
however with

- BenchmarkNodeHTMLBuildMany : doing `HTMLBuilder(&sb); HTMLBuilder(&sb); HTMLBuilder(&sb)`
- BenchmarkNodeHTMLMany: doing HTML() + HTML() + HTML()

`HTMLBuilder(sb *strings.Builder) *strings.Builder` doesn't look like it outperforms by much

```
BenchmarkAttrHTML-8            	 9900267	       121.1 ns/op	   9900267 AttrHTML	     128 B/op	       4 allocs/op
BenchmarkNodeHTML-8            	 2326978	       514.4 ns/op	   2326978 NodeHTML	     456 B/op	      11 allocs/op
BenchmarkNodeHTMLBuildMany-8   	  790669	      1490 ns/op	    790669 NodeHTML	    1640 B/op	      25 allocs/op
BenchmarkNodeHTMLMany-8        	  744154	      1591 ns/op	    744154 NodeHTML	    1656 B/op	      34 allocs/op
BenchmarkHtmlTemplate-8        	  357118	      3330 ns/op	    357118 HTMLTemplate	    1384 B/op	      53 allocs/op
```