@yongchongye 2023-11-10
then I close this PR? maybe relook at it if we hit any performance issues during load tests?